add_symbols_versioning
This patch adds a very simple linker script to version the lib's symbols
and thus trying to avoid crashes if an application loads two different
LLVM versions (as long as they do not share data between them).

Note that we deliberately *don't* make LLVM_3.9 depend on LLVM_3.8:
they're incompatible and the whole point of this patch is
to tell the linker that.

Author: Rebecca N. Palmer <rebecca_palmer@zoho.com>
Author: Lisandro Damían Nicanor Pérez Meyer <lisandro@debian.org>
Bug-Debian: https://bugs.debian.org/848368


Gbp-Pq: Name add_symbols_versioning.patch